Prefer readers in state locks
The normal behaviour for rw locks is to prefer readers, which can access
the protected resource concurrently. Use this mode of operation also for
cStateLock. So frequent read operations complete faster.

With this patch especially the VDR main thread is no longer blocked by
section handler threads, which was observed to happen up to several seconds.
This in turn delayed opening OSD menus and blocked handling of remote
control key presses for this amount of time, a very bad user experience.

Section handler filters are already prepared for timeouts on state locks
(channel and schedules write locks). So there is no problem with this
priority change. Also write accesses to timer and recording lists are not
timing critical, so all 4 types of state locks can use the 'prefer readers'
kind of operation.

Signed-off-by: Soeren Moch <smoch@web.de>